Initial commit
This project was for the Full Stack Developer bootcamp I did and was for a lesson on managing state in ReactJS. The assignment was to create an app that allows the user to choose between two functions – a currency converter and a Three-card Monte game. The currency converter uses the current exchange rates from https://apilayer.com/ in order to convert currency between US Dollars, Euros, Pounds Sterling, and South African Rands. The Three-Card Monte game is a simple game of chance that gives feedback about the winning or losing streak.